use database json for email storage. Clean up env variable setup
