+++ garden_ai/utils.py
@@ -2,6 +2,8 @@ import logging
 import sys
 from inspect import Signature, signature
 from itertools import zip_longest
+import json
+import base64
 
 import beartype.door
 import requests
@@ -156,3 +158,16 @@ def requests_to_curl(response: requests.Response) -> str:
     else:
         data = request.body
     return f"curl -X {method} -H {headers} -d '{data}' {uri}"
+
+
+def extract_email_from_globus_jwt(jwt: str) -> str:
+    try:
+        _, payload_b64, _ = jwt.split(".")
+        payload = json.loads(base64.b64decode(payload_b64 + "===").decode("utf-8"))
+    except Exception as e:
+        raise Exception('Invalid JWT') from e
+    try:
+        email = payload['identity_set'][0]['email']
+    except KeyError as e:
+        raise Exception('JWT did not include user email') from e
+    return email
